spotify_data/requirements.txt
Flask==2.3.3
spotipy==2.23.0
mysql-connector-python==8.1.0
requests==2.31.0
flask-cors==4.0.0
